Improved /api/route.ts code structure
